"AU_LINUX_ANDROID_LA.BF64.1.2.1_RB2.05.00.02.081.002"
* tag 'AU_LINUX_ANDROID_LA.BF64.1.2.1_RB2.05.00.02.081.002': (27
        * commits)
  Change dataUsage from long to int
  Grey out merge button after it is clicked.
  Telephony: Fix compilation issues.
  Revert "Fix NPE in findInfoQueryComplete"
  Revert "Update PhoneAccount icon API (4/6)"
  Revert "Don't cancel incall notification unnecessarily"
  Revert "Resize InCallUI for small devices."
  Revert "Fix NPE in getCallProviderLabel"
  Revert "Stop ringer when selecting "respond via text"."
  Revert "Telecom API updates (3/6)"
  Revert "Change Log.wtf to Log.w to prevent rare crash."
  Revert "Add some more logging in InCallActivity and InCallPresenter"
  Revert "Improve outgoing call UI responsiveness"
  Revert "In Multi-SIM cases use the color of the SIM icon for InCallUI."
  Revert "Telephony: Fix compilation issues for conversion from long to int."
  Telephony: Fix compilation issues for conversion from long to int.
  In Multi-SIM cases use the color of the SIM icon for InCallUI.
  Improve outgoing call UI responsiveness
  Add some more logging in
  InCallActivity and InCallPresenter
  Change Log.wtf to Log.w to prevent rare crash.
  ...

+++ src/com/android/incallui/AnswerPresenter.java
@@ -50,7 +50,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
             processVideoUpgradeRequestCall(call);
         }
         for (int i = 0; i < CallList.PHONE_COUNT; i++) {
-            long[] subId = CallList.getInstance().getSubId(i);
+            int[] subId = CallList.getInstance().getSubId(i);
             call = calls.getCallWithState(Call.State.INCOMING, 0, subId[0]);
             if (call == null) {
                 call = calls.getCallWithState(Call.State.CALL_WAITING, 0, subId[0]);
@@ -94,7 +94,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
 
     @Override
     public void onIncomingCall(Call call) {
-        long subId = call.getSubId();
+        int subId = call.getSubId();
         int phoneId = CallList.getInstance().getPhoneId(subId);
         // TODO: Ui is being destroyed when the fragment detaches.  Need clean up step to stop
         // getting updates here.
@@ -144,7 +144,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     }
 
     private void processIncomingCall(Call call) {
-        long subId = call.getSubId();
+        int subId = call.getSubId();
         int phoneId = CallList.getInstance().getPhoneId(subId);
         mCallId[phoneId] = call.getId();
         mCall[phoneId] = call;
@@ -168,7 +168,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
             return;
         }
 
-        long subId = call.getSubId();
+        int subId = call.getSubId();
         int phoneId = CallList.getInstance().getPhoneId(subId);
         mCallId[phoneId] = call.getId();
         mCall[phoneId] = call;
@@ -212,7 +212,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     public void onCallChanged(Call call) {
         Log.d(this, "onCallStateChange() " + call + " " + this);
         if (call.getState() != Call.State.INCOMING) {
-            long subId = call.getSubId();
+            int subId = call.getSubId();
             int phoneId = CallList.getInstance().getPhoneId(subId);
 
             boolean isUpgradePending = isVideoUpgradePending(call);
@@ -244,7 +244,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     private int getActivePhoneId() {
         int phoneId = -1;
         if (CallList.getInstance().isDsdaEnabled()) {
-            long subId = CallList.getInstance().getActiveSubscription();
+            int subId = CallList.getInstance().getActiveSubscription();
             phoneId = CallList.getInstance().getPhoneId(subId);
         } else {
             for (int i = 0; i < mCall.length; i++) {
@@ -370,7 +370,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     }
 
     @Override
-    public void onActiveSubChanged(long subId) {
+    public void onActiveSubChanged(int subId) {
         final CallList calls = CallList.getInstance();
         final Call call = calls.getIncomingCall();
         int phoneId = CallList.getInstance().getPhoneId(subId);
